# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Progi CAP API",
        "description": "API documentation for Progi CAP application modules",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/files/{file}/permissions": {
            "get": {
                "tags": [
                    "File Permissions"
                ],
                "summary": "Permissions d'un fichier",
                "description": "Récupère toutes les permissions accordées sur un fichier",
                "operationId": "getFilePermissions",
                "parameters": [
                    {
                        "name": "file",
                        "in": "path",
                        "description": "ID du fichier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permissions récupérées avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/FilePermission"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Fichier non trouvé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/files/{file}/permissions/grant": {
            "post": {
                "tags": [
                    "File Permissions"
                ],
                "summary": "Accorder une permission",
                "description": "Accorde une permission sur un fichier à un utilisateur ou un rôle",
                "operationId": "grantFilePermission",
                "parameters": [
                    {
                        "name": "file",
                        "in": "path",
                        "description": "ID du fichier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "permission_type"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID de l'utilisateur (optionnel si role_id fourni)",
                                        "type": "integer"
                                    },
                                    "role_id": {
                                        "description": "ID du rôle (optionnel si user_id fourni)",
                                        "type": "integer"
                                    },
                                    "permission_type": {
                                        "type": "string",
                                        "enum": [
                                            "read",
                                            "write",
                                            "delete",
                                            "share",
                                            "admin"
                                        ]
                                    },
                                    "expires_at": {
                                        "description": "Date d'expiration (optionnel)",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Permission accordée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Permission accordée avec succès."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/FilePermission"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Fichier non trouvé"
                    },
                    "422": {
                        "description": "Données invalides"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/files/{file}/permissions/revoke": {
            "post": {
                "tags": [
                    "File Permissions"
                ],
                "summary": "Révoquer une permission",
                "description": "Révoque une permission sur un fichier d'un utilisateur ou d'un rôle",
                "operationId": "revokeFilePermission",
                "parameters": [
                    {
                        "name": "file",
                        "in": "path",
                        "description": "ID du fichier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "permission_type"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "ID de l'utilisateur (optionnel si role_id fourni)",
                                        "type": "integer"
                                    },
                                    "role_id": {
                                        "description": "ID du rôle (optionnel si user_id fourni)",
                                        "type": "integer"
                                    },
                                    "permission_type": {
                                        "type": "string",
                                        "enum": [
                                            "read",
                                            "write",
                                            "delete",
                                            "share",
                                            "admin"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permission révoquée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Permission révoquée avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Fichier ou permission non trouvé"
                    },
                    "422": {
                        "description": "Données invalides"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/files/{file}/permissions/check": {
            "post": {
                "tags": [
                    "File Permissions"
                ],
                "summary": "Vérifier une permission",
                "description": "Vérifie si l'utilisateur actuel a une permission spécifique sur un fichier",
                "operationId": "checkFilePermission",
                "parameters": [
                    {
                        "name": "file",
                        "in": "path",
                        "description": "ID du fichier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "permission_type"
                                ],
                                "properties": {
                                    "permission_type": {
                                        "type": "string",
                                        "enum": [
                                            "read",
                                            "write",
                                            "delete",
                                            "share",
                                            "admin"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vérification effectuée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "has_permission": {
                                                    "type": "boolean"
                                                },
                                                "permission_type": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Fichier non trouvé"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "File": {
                "title": "File",
                "description": "Modèle représentant un fichier stocké",
                "properties": {
                    "id": {
                        "description": "ID unique",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "ID du propriétaire",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nom du fichier",
                        "type": "string"
                    },
                    "original_name": {
                        "description": "Nom original du fichier",
                        "type": "string"
                    },
                    "path": {
                        "description": "Chemin du fichier",
                        "type": "string"
                    },
                    "disk": {
                        "description": "Disque de stockage",
                        "type": "string"
                    },
                    "visibility": {
                        "description": "Visibilité du fichier",
                        "type": "string",
                        "enum": [
                            "public",
                            "private"
                        ]
                    },
                    "module_name": {
                        "description": "Nom du module",
                        "type": "string"
                    },
                    "module_resource_type": {
                        "description": "Type de ressource du module",
                        "type": "string"
                    },
                    "module_resource_id": {
                        "description": "ID de la ressource du module",
                        "type": "integer"
                    },
                    "collection": {
                        "description": "Collection du fichier",
                        "type": "string"
                    },
                    "size": {
                        "description": "Taille du fichier en octets",
                        "type": "integer"
                    },
                    "mime_type": {
                        "description": "Type MIME",
                        "type": "string"
                    },
                    "extension": {
                        "description": "Extension du fichier",
                        "type": "string"
                    },
                    "file_hash": {
                        "description": "Hash du fichier",
                        "type": "string"
                    },
                    "metadata": {
                        "description": "Métadonnées supplémentaires",
                        "type": "object"
                    },
                    "is_locked": {
                        "description": "Si le fichier est verrouillé",
                        "type": "boolean"
                    },
                    "locked_at": {
                        "description": "Date de verrouillage",
                        "type": "string",
                        "format": "date-time"
                    },
                    "locked_by": {
                        "description": "ID de l'utilisateur qui a verrouillé",
                        "type": "integer"
                    },
                    "download_count": {
                        "description": "Nombre de téléchargements",
                        "type": "integer"
                    },
                    "last_accessed_at": {
                        "description": "Dernier accès",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "date-time"
                    },
                    "url": {
                        "description": "URL du fichier",
                        "type": "string"
                    },
                    "size_for_humans": {
                        "description": "Taille formatée",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "FilePermission": {
                "title": "File Permission",
                "description": "Modèle représentant une permission sur un fichier",
                "properties": {
                    "id": {
                        "description": "ID unique",
                        "type": "integer"
                    },
                    "file_id": {
                        "description": "ID du fichier",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "ID de l'utilisateur (null si permission par rôle)",
                        "type": "integer",
                        "nullable": true
                    },
                    "role_id": {
                        "description": "ID du rôle (null si permission individuelle)",
                        "type": "integer",
                        "nullable": true
                    },
                    "permission_type": {
                        "description": "Type de permission",
                        "type": "string",
                        "enum": [
                            "read",
                            "write",
                            "delete",
                            "share",
                            "admin"
                        ]
                    },
                    "granted_by": {
                        "description": "ID de l'utilisateur qui a accordé la permission",
                        "type": "integer"
                    },
                    "granted_at": {
                        "description": "Date d'attribution",
                        "type": "string",
                        "format": "date-time"
                    },
                    "expires_at": {
                        "description": "Date d'expiration",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "date-time"
                    },
                    "file": {
                        "$ref": "#/components/schemas/File"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "role": {
                        "$ref": "#/components/schemas/Role"
                    }
                },
                "type": "object"
            },
...
            "Role": {
                "title": "Role",
                "description": "Modèle représentant un rôle utilisateur",
                "properties": {
                    "id": {
                        "description": "ID unique",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Nom du rôle",
                        "type": "string"
                    },
                    "display_name": {
                        "description": "Nom d'affichage",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description du rôle",
                        "type": "string"
                    },
                    "is_system": {
                        "description": "Si c'est un rôle système",
                        "type": "boolean"
                    },
                    "created_at": {
                        "description": "Date de création",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date de mise à jour",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}